Major update to iron_cuke internals
All the Gorillib stuff is now removed. It was confusing
and made it hard to add new tests to the framework. The
new system works as follows:

'Proctors' mediate the rendering of 'templates', which
are stored in lib/iron_cuke/templates. To render a new
test, simply create a new template and drop it in - it
should work automatically.

If you do not provide a proctor for you template, the
default proctor will be used instead. The default proctor
takest all keys and values from the announcements hash
and makes them instance variables, so if your test renders
rather straightforwardly from the announcements hash then
you can skip that step.

Otherwise you can define a proctor that does additional
processing or formatting. Currently, port and daemon proctors
are defined but not implemented.

The general idea is to be as close to the Rails view/controller
paradigm as possible, where proctors are the controllers and the
test templates are the views.